requirements.txt
dash==2.12.0
dash-core-components==2.0.0
dash-html-components==2.0.0
plotly==5.3.1
requests==2.26.0
 python-dotenv